API update to v2.1.0 for brick and cylindrical gauge chart.
