+++ src/CustomVisuals/Published/BrickChart/.api/v2.1.0/schema.dependencies.json
@@ -1,5 +1,5 @@
 {
-    "PBI_API_VERSION": "v1.4.0",
+    "PBI_API_VERSION": "v2.1.0",
     "type": "object",
     "properties": {
         "cranPackages": {
