+++ src/CustomVisuals/Published/BrickChart/.api/v2.1.0/schema.pbiviz.json
@@ -1,5 +1,5 @@
 {
-    "PBI_API_VERSION": "v1.4.0",
+    "PBI_API_VERSION": "v2.1.0",
     "type": "object",
     "properties": {
         "apiVersion": {
@@ -36,7 +36,15 @@
             "items": {
                 "type": "string"
             }
-        },        
+        },
+        "stringResources": {
+            "type": "array",
+            "description": "An array of relative paths to string resources to load",
+            "items": {
+                "type": "string"
+            },
+            "uniqueItems": true
+        },         
         "style" : {
             "type": "string",
             "description": "Relative path to the stylesheet (less) for the visual"
