+++ src/CustomVisuals/Published/BrickChart/package.json
@@ -1,17 +1,17 @@
 {
   "name": "visual",
   "dependencies": {
+    "async": "^0.1.22",
     "d3": "^3.5.5",
+    "globalize": "0.1.0-a2",
+    "jquery": "^3.3.1",
+    "lodash": "^4.17.10",
+    "powerbi-visuals-utils-chartutils": "^0.3.0",
     "powerbi-visuals-utils-dataviewutils": "^1.0.1",
     "powerbi-visuals-utils-formattingutils": "^0.2.2",
-    "powerbi-visuals-utils-tooltiputils": "^0.3.0",
-    "powerbi-visuals-utils-chartutils": "^0.3.0",
-    "powerbi-visuals-utils-svgutils": "^0.2.1",
-    "powerbi-visuals-utils-typeutils": "^0.2.1",
     "powerbi-visuals-utils-interactivityutils": "^0.2.1",
-    "jquery": "^3.1.1",
-    "async": "^0.1.22",
-    "lodash": "^4.16.2",
-    "globalize": "0.1.0-a2"
+    "powerbi-visuals-utils-svgutils": "^0.2.1",
+    "powerbi-visuals-utils-tooltiputils": "^0.3.2",
+    "powerbi-visuals-utils-typeutils": "^0.2.1"
   }
 }
