+++ src/CustomVisuals/Published/BrickChart/pbiviz.json
@@ -4,12 +4,12 @@
     "displayName": "Brick Chart by MAQ Software",
     "guid": "BrickChart1454871868044",
     "visualClassName": "BrickChart",
-    "version": "3.1.2",
-    "description": "Brick Chart consists of 100 squares that are colored according to the percentage breakdown of your datasets. Hover your mouse over a square to bring up a tooltip. The tooltip indicates which dataset the color represents and the percentage value of that category. An optional legend above the chart identifies which datasets correspond with which colors. You may tailor the legend’s title, size and color. You may also customize the chart’s width and height.",
+    "version": "3.1.3",
+    "description": "Brick Chart by MAQ Software allows users to display a percentage breakdown of a category in an alternative format to a donut or pie chart. This visual is useful for showing percentages or positions occupied, such as seats in a legislature. It is also useful for comparing each portion of a category to the whole. It can be viewed either in a square or rectangular format with options to have different types of brick shapes. Animations available for visual effects.  ",
     "supportUrl": "https://github.com/maqsoftware/PowerBI-visuals/issues",
     "gitHubUrl": "https://github.com/maqsoftware/PowerBI-visuals/"
   },
-  "apiVersion": "1.8.0",
+  "apiVersion": "2.1.0",
   "author": {
     "name": "MAQ Software",
     "email": "support@maqsoftware.com"
@@ -31,7 +31,8 @@
     "node_modules/powerbi-visuals-utils-svgutils/lib/index.js",
     "node_modules/powerbi-visuals-utils-formattingutils/lib/index.js",
     "node_modules/powerbi-visuals-utils-interactivityutils/lib/index.js",
-    "node_modules/powerbi-visuals-utils-chartutils/lib/index.js"
+    "node_modules/powerbi-visuals-utils-chartutils/lib/index.js",
+    "node_modules/powerbi-visuals-utils-tooltiputils/lib/index.js"
   ],
   "style": "style/visual.less",
   "capabilities": "capabilities.json",
