+++ src/CustomVisuals/Published/BrickChart/style/visual.less
@@ -1,7 +1,3 @@
-.brickchart_topContainer .svg {
-    width: 211px;
-    height: 211px;
-}
 .brickchart_topContainer{
     position: absolute;
     line-height: normal;
@@ -14,6 +10,23 @@
     margin-top: 0;
     margin-bottom: 0;
     transform-origin: 50% 0 0;
-	height: 211px;
-	width : 211px;
 }
+.bc_ErrorMessage{
+    text-align: center;
+    font-size: 20px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    top: 45%;
+    width:100%;
+    position: absolute;
+}
+.bc_ErrorMsg{
+    text-align: center;
+    font-size: 20px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    top: 45%;
+    width:100%;
+    position: absolute;
+}
+
