+++ src/CustomVisuals/Published/BrickChart/tsconfig.json
@@ -8,8 +8,7 @@
     "out": "./.tmp/build/visual.js"
   },
   "files": [
-    ".api/v1.8.0/PowerBI-visuals.d.ts",
-    "typings/index.d.ts",
+    ".api/v2.1.0/PowerBI-visuals.d.ts",
     "src/visual.ts",
     "node_modules/powerbi-visuals-utils-dataviewutils/lib/index.d.ts",
     "typings/index.d.ts",
@@ -18,7 +17,6 @@
     "node_modules/powerbi-visuals-utils-svgutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-typeutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-chartutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-tooltiputils/lib/index.d.ts",
-    "src/tooltipServiceWrapper.ts"
+    "node_modules/powerbi-visuals-utils-tooltiputils/lib/index.d.ts"
   ]
 }
