+++ src/CustomVisuals/Published/CylindricalGauge/pbiviz.json
@@ -4,15 +4,15 @@
     "displayName": "Cylindrical Gauge by MAQ Software",
     "guid": "cylindricalGauge73A12A442345453EB69B593649C3A341",
     "visualClassName": "CylindricalGauge",
-    "version": "3.3.1",
+    "version": "3.3.2",
     "description": "Track current values against a target with an intuitive 3D cylinder visual",
     "supportUrl": "https://github.com/maqsoftware/PowerBI-visuals/issues",
     "gitHubUrl": "https://github.com/maqsoftware/PowerBI-visuals/"
   },
-  "apiVersion": "1.4.0",
+  "apiVersion": "2.1.0",
   "author": {
     "name": "MAQ Software",
-    "email": "sales@maqsoftware.com"
+    "email": "support@maqsoftware.com"
   },
   "assets": {
     "icon": "./assets/icon.png"
