+++ src/CustomVisuals/Published/CylindricalGauge/tsconfig.json
@@ -8,12 +8,12 @@
     "out": "./.tmp/build/visual.js"
   },
   "files": [
-    ".api/v1.4.0/PowerBI-visuals.d.ts",
+    ".api/v2.1.0/PowerBI-visuals.d.ts",
     "typings/index.d.ts",
     "node_modules/powerbi-visuals-utils-typeutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-svgutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-dataviewutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-formattingutils/lib/index.d.ts",  
+    "node_modules/powerbi-visuals-utils-formattingutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-interactivityutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-chartutils/lib/index.d.ts",
     "src/visual.ts",
